fix(SFINT-3713): Ensure types match with module
What's the issue:
When importing CA.js as an ESM, the typings getting resolved are currently `library.d.ts`, however in the `rollup.config.js` we can see that `library.es.js` is built with `headless.ts` as its entry point, meaning that you end up with a mismatch between what you import and what you get. I do not know if that's what we want, so I open a PR to at least discuss it.